intermediate added tests
