+++ src/databricks/labs/remorph/assessments/pipeline.py
@@ -1,6 +1,7 @@
 import logging
 import yaml
 import duckdb
+import pandas as pd
 
 from databricks.labs.remorph.assessments.profiler_config import PipelineConfig, Step
 from databricks.labs.remorph.connections.database_manager import DatabaseManager
@@ -37,16 +38,19 @@ class PipelineClass:
     def _save_to_db(self, result, step_name: str, mode: str):
         conn = duckdb.connect('pipeline_results.duckdb')
         cursor = conn.cursor()
+        df = pd.DataFrame(result.fetchall(), columns=result.keys())
+        print(df.head(5))
+
 
         # Save result to Parquet file
-        parquet_file = f"{step_name}.parquet"
         if mode == "overwrite":
             logging.debug(f"Overwriting table: {step_name}")
             cursor.execute(f"DROP TABLE IF EXISTS {step_name}")
-            cursor.execute(f"COPY ({result}) TO '{parquet_file}' (FORMAT PARQUET)")
+            cursor.execute(f"CREATE TABLE {step_name} AS SELECT * FROM {df}")
         elif mode == "append":
             logging.debug(f"Appending to table: {step_name}")
-            cursor.execute(f"COPY ({result}) TO '{parquet_file}' (FORMAT PARQUET)")
+            cursor.execute(f"CREATE TABLE IF NOT EXISTS {step_name} AS {df} LIMIT 0")
+            cursor.execute(f"INSERT INTO {step_name} SELECT * FROM {df}")
 
         conn.close()
