+++ tests/resources/assessments/inventory.sql
@@ -1,0 +1,6 @@
+select DB_ID(name) as db_id,
+       name,
+       collation_name,
+       create_date,
+       SYSDATETIME() as extract_ts
+from SYS.DATABASES
