+++ tests/resources/assessments/pipeline_config.yml
@@ -1,0 +1,21 @@
+name: ExamplePipeline
+version: "1.0"
+steps:
+  - name: inventory
+    type: inventory
+    extract_query: resources/assessments/inventory.sql
+    mode: overwrite
+    frequency: daily
+    flag: active
+  - name: usage
+    type: usage
+    extract_query: resources/assessments/usage.sql
+    mode: overwrite
+    frequency: weekly
+    flag: active
+  - name: usage_2
+    type: usage_2
+    extract_query: resources/assessments/usage.sql
+    mode: overwrite
+    frequency: daily
+    flag: inactive
